Convert Midi levels (vol/pan/pitch) to use byte values until conversion to linear is needed

+++ src/com/kh/beatbot/midi/event/NoteOff.java
@@ -18,12 +18,11 @@ package com.kh.beatbot.midi.event;
 
 public class NoteOff extends ChannelEvent {
 
-	public NoteOff(long tick, int channel, int note, float velocity, float pan, float pitch) {
+	public NoteOff(long tick, int channel, int note, byte velocity, byte pan, byte pitch) {
 		super(tick, ChannelEvent.NOTE_OFF, channel, note, velocity, pan, pitch);
 	}
 
-	public NoteOff(long tick, long delta, int channel, int note, float velocity, float pan,
-			float pitch) {
+	public NoteOff(long tick, long delta, int channel, int note, byte velocity, byte pan, byte pitch) {
 		super(tick, delta, ChannelEvent.NOTE_OFF, channel, note, velocity, pan, pitch);
 	}
 }
